requirements.txt
flask
python-dotenv
huggingface_hub
